Start to add MediaPlayerCarousel
