+++ package.json
@@ -65,6 +65,7 @@
     "react-native-config": "1.4.1",
     "react-native-device-info": "7.3.1",
     "react-native-dialog": "5.6.0",
+    "react-native-dots-pagination": "^0.1.11",
     "react-native-elements": "1.2.7",
     "react-native-fast-image": "7.0.2",
     "react-native-fs": "2.16.6",
