+++ src/components/index.ts
@@ -10,6 +10,10 @@ export { HeaderTitleSelector } from './HeaderTitleSelector'
 export { HTMLScrollView } from './HTMLScrollView'
 export { IndicatorDownload } from './IndicatorDownload'
 export { Login } from './Login'
+export { MediaPlayerCarousel } from './MediaPlayerCarousel'
+export { MediaPlayerCarouselClips } from './MediaPlayerCarouselClips'
+export { MediaPlayerCarouselShowNotes } from './MediaPlayerCarouselShowNotes'
+export { MediaPlayerCarouselViewer } from './MediaPlayerCarouselViewer'
 export { MessageWithAction } from './MessageWithAction'
 export { MoreButton } from './MoreButton'
 export { NavAddToPlaylistIcon } from './NavAddToPlaylistIcon'
