+++ src/screens/PlayerScreen.tsx
@@ -1,20 +1,11 @@
-import {
-  convertNowPlayingItemToEpisode,
-  convertNowPlayingItemToMediaRef,
-  convertToNowPlayingItem,
-  NowPlayingItem
-} from 'podverse-shared'
+import { convertNowPlayingItemToMediaRef, convertToNowPlayingItem } from 'podverse-shared'
 import { StyleSheet, View as RNView } from 'react-native'
 import Share from 'react-native-share'
 import React, { getGlobal, setGlobal } from 'reactn'
 import {
   ActionSheet,
-  ActivityIndicator,
   ClipInfoView,
-  ClipTableCell,
-  Divider,
-  FlatList,
-  HTMLScrollView,
+  MediaPlayerCarousel,
   NavAddToPlaylistIcon,
   NavDismissIcon,
   NavMakeClipIcon,
@@ -24,22 +15,14 @@ import {
   PlayerClipInfoBar,
   PlayerControls,
   PlayerTableHeader,
-  TableSectionSelectors,
   View
 } from '../components'
-import { downloadEpisode } from '../lib/downloader'
 import { translate } from '../lib/i18n'
 import { hasValidNetworkConnection } from '../lib/network'
-import {
-  formatTitleViewHtml,
-  readableDate,
-  replaceLinebreaksWithBrTags,
-  safelyUnwrapNestedVariable,
-  testProps
-} from '../lib/utility'
+import { replaceLinebreaksWithBrTags, safelyUnwrapNestedVariable, testProps } from '../lib/utility'
 import { PV } from '../resources'
-import { getEpisode, retrieveLatestChaptersForEpisodeId } from '../services/episode'
-import { getMediaRef, getMediaRefs } from '../services/mediaRef'
+import { getEpisode } from '../services/episode'
+import { getMediaRef } from '../services/mediaRef'
 import { PVTrackPlayer } from '../services/player'
 import PlayerEventEmitter from '../services/playerEventEmitter'
 import { addQueueItemNext } from '../services/queue'
@@ -57,15 +40,6 @@ type State = {}
 const testIDPrefix = 'player_screen'
 
 let eventListenerPlayerNewEpisodeLoaded: any
-
-/* 
-  The shouldQueryAgain variable is used to determine if the PlayerScreen should reload its data
-  on componentDidMount. This is (*I think*) intended to handle the condition where the app is returning
-  to the foreground from the background, which will trigger another componentDidMount, we will want
-  to refresh the screen's data.
-  shouldQueryAgain is set to true when the PLAYER_NEW_EPISODE_LOADED event is emitted.
-*/
-let shouldQueryAgain = false
 
 export class PlayerScreen extends React.Component<Props, State> {
   static navigationOptions = ({ navigation }) => {
@@ -124,11 +98,6 @@ export class PlayerScreen extends React.Component<Props, State> {
       _showShareActionSheet: this._showShareActionSheet
     })
 
-    if (shouldQueryAgain) {
-      shouldQueryAgain = false
-      this._selectViewType(this.global.screenPlayer.viewType)
-    }
-
     if (!eventListenerPlayerNewEpisodeLoaded) {
       eventListenerPlayerNewEpisodeLoaded = PlayerEventEmitter.on(
         PV.Events.PLAYER_NEW_EPISODE_LOADED,
@@ -142,7 +111,6 @@ export class PlayerScreen extends React.Component<Props, State> {
   }
 
   _handleNewEpisodeLoaded = async () => {
-    shouldQueryAgain = true
     setTimeout(() => {
       this._handleUpdateFullEpisode()
     }, 5000)
@@ -233,203 +201,11 @@ export class PlayerScreen extends React.Component<Props, State> {
     }
   }
 
-  _selectViewType = async (selectedKey: string) => {
-    if (!selectedKey) {
-      setGlobal({
-        screenPlayer: {
-          ...this.global.screenPlayer,
-          viewType: null
-        }
-      })
-      return
-    }
-
-    let sort = PV.Filters._topPastWeek
-    let hideRightItemWhileLoading = false
-    if (selectedKey === PV.Filters._clipsKey) {
-      sort = PV.Filters._chronologicalKey
-      hideRightItemWhileLoading = true
-    }
-
-    setGlobal(
-      {
-        screenPlayer: {
-          ...this.global.screenPlayer,
-          endOfResultsReached: false,
-          flatListData: [],
-          flatListDataTotalCount: null,
-          hideRightItemWhileLoading,
-          isQuerying: true,
-          queryFrom: PV.Filters._fromThisEpisodeKey,
-          querySort: sort,
-          queryPage: 1,
-          viewType: selectedKey
-        }
-      },
-      async () => {
-        if (selectedKey === PV.Filters._chaptersKey || selectedKey === PV.Filters._clipsKey) {
-          const newState = await this._queryData()
-          setGlobal({
-            screenPlayer: {
-              ...this.global.screenPlayer,
-              ...newState
-            }
-          })
-        } else {
-          setGlobal({
-            screenPlayer: {
-              ...this.global.screenPlayer,
-              isQuerying: false
-            }
-          })
-        }
-      }
-    )
-  }
-
-  _selectQueryFrom = async (selectedKey: string) => {
-    if (!selectedKey) {
-      setGlobal({
-        screenPlayer: {
-          ...this.global.screenPlayer,
-          queryFrom: null
-        }
-      })
-      return
-    }
-
-    let sort = PV.Filters._topPastWeek
-    let hideRightItemWhileLoading = false
-    if (selectedKey === PV.Filters._fromThisEpisodeKey) {
-      sort = PV.Filters._chronologicalKey
-      hideRightItemWhileLoading = true
-    }
-
-    setGlobal(
-      {
-        screenPlayer: {
-          ...this.global.screenPlayer,
-          endOfResultsReached: false,
-          flatListData: [],
-          flatListDataTotalCount: null,
-          hideRightItemWhileLoading,
-          isQuerying: true,
-          queryFrom: selectedKey,
-          queryPage: 1,
-          querySort: sort
-        }
-      },
-      async () => {
-        const newState = await this._queryData()
-        setGlobal({
-          screenPlayer: {
-            ...this.global.screenPlayer,
-            ...newState
-          }
-        })
-      }
-    )
-  }
-
-  _selectQuerySort = async (selectedKey: string) => {
-    if (!selectedKey) {
-      setGlobal({
-        screenPlayer: {
-          ...this.global.screenPlayer,
-          querySort: null
-        }
-      })
-      return
-    }
-
-    setGlobal(
-      {
-        screenPlayer: {
-          ...this.global.screenPlayer,
-          endOfResultsReached: false,
-          flatListData: [],
-          flatListDataTotalCount: null,
-          isQuerying: true,
-          querySort: selectedKey
-        }
-      },
-      async () => {
-        const newState = await this._queryData()
-        setGlobal({
-          screenPlayer: {
-            ...this.global.screenPlayer,
-            ...newState
-          }
-        })
-      }
-    )
-  }
-
-  _onEndReached = ({ distanceFromEnd }) => {
-    const { screenPlayer } = this.global
-    const { endOfResultsReached, isLoadingMore, queryPage = 1, viewType } = screenPlayer
-    if (
-      viewType !== PV.Filters._showNotesKey &&
-      viewType !== PV.Filters._titleKey &&
-      !endOfResultsReached &&
-      !isLoadingMore
-    ) {
-      if (distanceFromEnd > -1) {
-        setGlobal(
-          {
-            screenPlayer: {
-              ...this.global.screenPlayer,
-              isLoadingMore: true,
-              queryPage: queryPage + 1
-            }
-          },
-          async () => {
-            const newState = await this._queryData()
-            setGlobal({
-              screenPlayer: {
-                ...this.global.screenPlayer,
-                ...newState
-              }
-            })
-          }
-        )
-      }
-    }
-  }
-
-  _ItemSeparatorComponent = () => {
-    return <Divider />
-  }
-
   _toggleShowFullClipInfo = () => {
     setGlobal({
       screenPlayer: {
         ...this.global.screenPlayer,
         showFullClipInfo: !this.global.screenPlayer.showFullClipInfo
-      }
-    })
-  }
-
-  _handleMoreCancelPress = () => {
-    return new Promise((resolve, reject) => {
-      setGlobal(
-        {
-          screenPlayer: {
-            ...this.global.screenPlayer,
-            showMoreActionSheet: false
-          }
-        },
-        resolve
-      )
-    })
-  }
-
-  _handleMorePress = (selectedItem: any) => {
-    setGlobal({
-      screenPlayer: {
-        ...this.global.screenPlayer,
-        selectedItem,
-        showMoreActionSheet: true
       }
     })
   }
@@ -483,77 +259,11 @@ export class PlayerScreen extends React.Component<Props, State> {
     this._dismissShareActionSheet()
   }
 
-  _handleNavigationPress = (selectedItem: any) => {
-    const shouldPlay = true
-    loadItemAndPlayTrack(selectedItem, shouldPlay)
-  }
-
-  _handleDownloadPressed = () => {
-    const { selectedItem } = this.global.screenPlayer
-    if (selectedItem) {
-      const episode = convertNowPlayingItemToEpisode(selectedItem)
-      downloadEpisode(episode, episode.podcast)
-    }
-  }
-
-  _renderItem = ({ item, index }) => {
-    const { player, screenPlayer } = this.global
-    const { episode } = player
-    const podcast = (episode && episode.podcast) || {}
-    const { queryFrom } = screenPlayer
-
-    if (queryFrom === PV.Filters._fromThisEpisodeKey) {
-      item = {
-        ...item,
-        episode
-      }
-    }
-
-    return item && item.episode && item.episode.id ? (
-      <ClipTableCell
-        endTime={item.endTime}
-        episodeId={item.episode.id}
-        {...(queryFrom === PV.Filters._fromThisPodcastKey
-          ? { episodePubDate: readableDate(item.episode.pubDate) }
-          : {})}
-        {...(queryFrom === PV.Filters._fromThisPodcastKey
-          ? { episodeTitle: item.episode.title || translate('untitled episode') }
-          : {})}
-        handleMorePress={() => this._handleMorePress(convertToNowPlayingItem(item, null, podcast))}
-        handleNavigationPress={() => this._handleNavigationPress(convertToNowPlayingItem(item, null, podcast))}
-        hideImage={true}
-        showEpisodeInfo={queryFrom !== PV.Filters._fromThisEpisodeKey}
-        showPodcastTitle={false}
-        startTime={item.startTime}
-        testID={`${testIDPrefix}_clip_item_${index}`}
-        {...(item.title ? { title: item.title } : {})}
-        transparent={true}
-      />
-    ) : (
-      <></>
-    )
-  }
-
   render() {
     const { navigation } = this.props
-    const { offlineModeEnabled, player, screenPlayer } = this.global
+    const { player, screenPlayer } = this.global
     const { episode, nowPlayingItem } = player
-    const {
-      flatListData,
-      flatListDataTotalCount,
-      hideRightItemWhileLoading,
-      isLoading,
-      isLoadingMore,
-      isQuerying,
-      queryFrom,
-      querySort,
-      selectedItem,
-      showFullClipInfo,
-      showMoreActionSheet,
-      showNoInternetConnectionMessage,
-      showShareActionSheet,
-      viewType
-    } = screenPlayer
+    const { isLoading, showFullClipInfo, showShareActionSheet, viewType } = screenPlayer
     let { mediaRef } = player
 
     if (nowPlayingItem && nowPlayingItem.clipId) {
@@ -568,209 +278,46 @@ export class PlayerScreen extends React.Component<Props, State> {
       episode.description = replaceLinebreaksWithBrTags(episode.description)
     }
 
-    let noResultsMessage = translate('No episodes found')
-    let noResultsSubMessage = ''
-    if (viewType === PV.Filters._chaptersKey) {
-      noResultsMessage = translate('No chapters found')
-      noResultsSubMessage = translate('Chapters are created by the podcaster')
-    } else if (viewType === PV.Filters._clipsKey) {
-      noResultsMessage = translate('No clips found')
-    }
-
-    const showOfflineMessage =
-      offlineModeEnabled && queryFrom !== PV.Filters._showNotesKey && queryFrom !== PV.Filters._titleKey
-
     return (
-      <OpaqueBackground nowPlayingItem={nowPlayingItem}>
-        <View style={styles.view} transparent={true} {...testProps('player_screen_view')}>
-          <PlayerTableHeader nowPlayingItem={nowPlayingItem} testID={testIDPrefix} />
-          {showFullClipInfo && (mediaRef || (nowPlayingItem && nowPlayingItem.clipId)) && (
-            <ClipInfoView
-              createdAt={mediaRef.createdAt}
-              endTime={mediaRef.endTime}
-              handleClosePress={this._toggleShowFullClipInfo}
-              hideDynamicAdsWarning={nowPlayingItem.podcastHideDynamicAdsWarning}
-              isLoading={isLoading}
-              isOfficialChapter={mediaRef.isOfficialChapter}
-              isOfficialSoundBite={mediaRef.isOfficialSoundBite}
-              isPublic={mediaRef.isPublic}
-              navigation={navigation}
-              {...(mediaRef.owner ? { ownerId: mediaRef.owner.id } : {})}
-              {...(mediaRef.owner ? { ownerIsPublic: mediaRef.owner.isPublic } : {})}
-              {...(mediaRef.owner ? { ownerName: mediaRef.owner.name } : {})}
-              startTime={mediaRef.startTime}
-              {...(mediaRef.title ? { title: mediaRef.title } : {})}
-            />
-          )}
-          {!showFullClipInfo && (
-            <View style={styles.view} transparent={true}>
-              <TableSectionSelectors
-                handleSelectLeftItem={this._selectViewType}
-                handleSelectRightItem={this._selectQuerySort}
-                hideRightItemWhileLoading={hideRightItemWhileLoading}
-                includeChronological={viewType === PV.Filters._clipsKey && queryFrom === PV.Filters._fromThisEpisodeKey}
-                isTransparent={true}
-                screenName='PlayerScreen'
-                selectedLeftItemKey={viewType}
-                selectedRightItemKey={querySort}
-                testID={testIDPrefix}
+      <React.Fragment>
+        <OpaqueBackground nowPlayingItem={nowPlayingItem}>
+          <View style={styles.view} transparent={true} {...testProps('player_screen_view')}>
+            <PlayerTableHeader nowPlayingItem={nowPlayingItem} testID={testIDPrefix} />
+            {!showFullClipInfo && <MediaPlayerCarousel />}
+            {showFullClipInfo && (mediaRef || (nowPlayingItem && nowPlayingItem.clipId)) && (
+              <ClipInfoView
+                createdAt={mediaRef.createdAt}
+                endTime={mediaRef.endTime}
+                handleClosePress={this._toggleShowFullClipInfo}
+                hideDynamicAdsWarning={nowPlayingItem.podcastHideDynamicAdsWarning}
+                isLoading={isLoading}
+                isOfficialChapter={mediaRef.isOfficialChapter}
+                isOfficialSoundBite={mediaRef.isOfficialSoundBite}
+                isPublic={mediaRef.isPublic}
+                navigation={navigation}
+                {...(mediaRef.owner ? { ownerId: mediaRef.owner.id } : {})}
+                {...(mediaRef.owner ? { ownerIsPublic: mediaRef.owner.isPublic } : {})}
+                {...(mediaRef.owner ? { ownerName: mediaRef.owner.name } : {})}
+                startTime={mediaRef.startTime}
+                {...(mediaRef.title ? { title: mediaRef.title } : {})}
               />
-              {viewType === PV.Filters._clipsKey && (
-                <TableSectionSelectors
-                  handleSelectLeftItem={this._selectQueryFrom}
-                  isBottomBar={true}
-                  isTransparent={true}
-                  screenName='PlayerScreen'
-                  selectedLeftItemKey={queryFrom}
-                  testID={`${testIDPrefix}_sub`}
-                />
-              )}
-              {isLoading || (isQuerying && <ActivityIndicator />)}
-              {!isLoading &&
-                !isQuerying &&
-                viewType &&
-                viewType !== PV.Filters._showNotesKey &&
-                viewType !== PV.Filters._titleKey &&
-                flatListData && (
-                  <FlatList
-                    data={flatListData}
-                    dataTotalCount={flatListDataTotalCount}
-                    disableLeftSwipe={true}
-                    extraData={flatListData}
-                    isLoadingMore={isLoadingMore}
-                    ItemSeparatorComponent={this._ItemSeparatorComponent}
-                    keyExtractor={(item: any) => item.id}
-                    noResultsMessage={noResultsMessage}
-                    noResultsSubMessage={noResultsSubMessage}
-                    onEndReached={this._onEndReached}
-                    renderItem={this._renderItem}
-                    showNoInternetConnectionMessage={showOfflineMessage || showNoInternetConnectionMessage}
-                    transparent={true}
-                  />
-                )}
-              {!isLoading && viewType === PV.Filters._showNotesKey && episode && (
-                <HTMLScrollView
-                  fontSizeLargestScale={PV.Fonts.largeSizes.md}
-                  html={episode.description ? episode.description : ''}
-                />
-              )}
-              {!isLoading && viewType === PV.Filters._titleKey && episode && (
-                <HTMLScrollView fontSizeLargestScale={PV.Fonts.largeSizes.md} html={formatTitleViewHtml(episode)} />
-              )}
-            </View>
-          )}
-          {nowPlayingItem && nowPlayingItem.clipId && (
-            <PlayerClipInfoBar handleOnPress={this._toggleShowFullClipInfo} nowPlayingItem={nowPlayingItem} />
-          )}
-          <PlayerControls navigation={navigation} />
-          <ActionSheet
-            handleCancelPress={this._handleMoreCancelPress}
-            items={() =>
-              PV.ActionSheet.media.moreButtons(selectedItem, navigation, {
-                handleDismiss: this._handleMoreCancelPress,
-                handleDownload: this._handleDownloadPressed
-              })
-            }
-            showModal={showMoreActionSheet}
-            testID={`${testIDPrefix}_more`}
-          />
-          <ActionSheet
-            handleCancelPress={this._dismissShareActionSheet}
-            items={shareActionSheetButtons(podcastId, episodeId, mediaRefId, this._handleShare)}
-            message={translate('What link do you want to share?')}
-            showModal={showShareActionSheet}
-            testID={`${testIDPrefix}_share`}
-            title={translate('Share')}
-          />
-        </View>
-      </OpaqueBackground>
+            )}
+            {nowPlayingItem && nowPlayingItem.clipId && (
+              <PlayerClipInfoBar handleOnPress={this._toggleShowFullClipInfo} nowPlayingItem={nowPlayingItem} />
+            )}
+            <PlayerControls navigation={navigation} />
+            <ActionSheet
+              handleCancelPress={this._dismissShareActionSheet}
+              items={shareActionSheetButtons(podcastId, episodeId, mediaRefId, this._handleShare)}
+              message={translate('What link do you want to share?')}
+              showModal={showShareActionSheet}
+              testID={`${testIDPrefix}_share`}
+              title={translate('Share')}
+            />
+          </View>
+        </OpaqueBackground>
+      </React.Fragment>
     )
-  }
-
-  _queryChapters = async () => {
-    const { player } = this.global
-    const { nowPlayingItem } = player
-
-    if (nowPlayingItem && !nowPlayingItem.addByRSSPodcastFeedUrl) {
-      return retrieveLatestChaptersForEpisodeId(nowPlayingItem.episodeId)
-    } else {
-      return [[], 0]
-    }
-  }
-
-  _queryClips = async () => {
-    const { player, screenPlayer } = this.global
-    const { nowPlayingItem } = player
-    const { queryFrom, queryPage } = screenPlayer
-
-    const sort = this._validSort()
-
-    if (nowPlayingItem && !nowPlayingItem.addByRSSPodcastFeedUrl) {
-      const results = await getMediaRefs({
-        sort,
-        page: queryPage,
-        ...(queryFrom === PV.Filters._fromThisEpisodeKey && nowPlayingItem
-          ? { episodeId: nowPlayingItem.episodeId }
-          : {}),
-        ...(queryFrom === PV.Filters._fromThisPodcastKey && nowPlayingItem
-          ? { podcastId: nowPlayingItem.podcastId }
-          : {}),
-        includeEpisode: queryFrom === PV.Filters._fromThisPodcastKey,
-        allowUntitled: true
-      })
-
-      return results
-    } else {
-      return [[], 0]
-    }
-  }
-
-  _validSort = () => {
-    const { screenPlayer } = this.global
-    const { queryFrom, querySort } = screenPlayer
-
-    return !querySort || (queryFrom === PV.Filters._fromThisPodcastKey && querySort === PV.Filters._chronologicalKey)
-      ? PV.Filters._topPastWeek
-      : querySort
-  }
-
-  _queryData = async (item?: NowPlayingItem, page?: number) => {
-    const { screenPlayer } = this.global
-    const { flatListData, viewType } = screenPlayer
-    const newState = {
-      hideRightItemWhileLoading: false,
-      isLoading: false,
-      isLoadingMore: false,
-      isQuerying: false,
-      showNoInternetConnectionMessage: false
-    } as any
-
-    const hasInternetConnection = await hasValidNetworkConnection()
-
-    if (!hasInternetConnection) {
-      newState.showNoInternetConnectionMessage = true
-      return newState
-    }
-
-    try {
-      if (viewType === PV.Filters._clipsKey) {
-        const results = await this._queryClips()
-        newState.flatListData = [...flatListData, ...results[0]]
-        newState.endOfResultsReached = newState.flatListData.length >= results[1]
-        newState.flatListDataTotalCount = results[1]
-      } else if (viewType === PV.Filters._chaptersKey) {
-        const results = await this._queryChapters()
-        newState.flatListData = [...flatListData, ...results[0]]
-        newState.endOfResultsReached = newState.flatListData.length >= results[1]
-        newState.flatListDataTotalCount = results[1]
-      }
-
-      newState.querySort = this._validSort()
-
-      return newState
-    } catch (error) {
-      return newState
-    }
   }
 }
